scored: one boolean per strategy
