+++ libs/iresearch/include/iresearch/CMakeLists.txt
@@ -340,11 +340,10 @@ set(IResearch_core_sources
     ./search/scored/make_posting.cpp
     ./search/scored/make_single_posting.cpp
     ./search/scored/make_all.cpp
-    ./search/scored/make_window_disjunction.cpp
-    ./search/scored/make_sparse_conjunction.cpp
+    ./search/scored/make_boolean.cpp
+    ./search/scored/make_boolean_window.cpp
+    ./search/scored/make_boolean_sparse.cpp
     ./search/scored/make_boosted_posting.cpp
-    ./search/scored/make_sparse_exclusion.cpp
-    ./search/scored/make_window_exclusion.cpp
     ./search/scored/make_bits_threshold.cpp
     ./search/scored/make_count_threshold.cpp
     ./search/scored/make_fixed_phrase.cpp
